fish-lights/requirements.txt
termcolor
requests
dropbox
RPi.GPIO
pytz
pika
